backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.12.0
chromadb==0.4.22
pdfplumber==0.10.4
sse-starlette==1.8.2
tiktoken==0.5.2
aiofiles==23.2.1
